+++ CHANGELOG.md
@@ -11,6 +11,10 @@ the code was deployed.
 
 ## [Unreleased]
 
+### Changed
+
+- Chatbot text messages no longer assume that the Sensor is installed in a bathroom (CU-271q62d).
+
 ## [5.2.0] - 2022-05-09
 
 ### Added